BIM: move add/removeSpaceBoundaries definitions to the Arch module
